pyproject.toml
[tool.poetry]
name = "fastapi-ml-template"
version = "0.1.0"
description = "A template for a production-ready machine learning REST API"
authors = ["Harrison Hoffman"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.2.1"
numpy = "^1.26.4"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.5"
black = "^24.3.0"
mypy = "^1.9.0"
pre-commit = "^3.7.0"

[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F", "I"]
fixable = ["ALL"]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true
plugins = "pydantic.mypy"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
